pyproject.toml
[tool.poetry]
name = "project_todo"
version = "0.1.0"
description = "A TODO list project made for computer science class"
authors = ["NicolasFloresF <nick.lasff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flet = "^0.23.0"
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
